creating buy good route

+++ backend/models/user.go
@@ -4,12 +4,12 @@ import "gorm.io/gorm"
 
 type User struct {
 	gorm.Model
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-	Email     string `json:"email"`
-	Age       string `json:"age"`
-	Phone     string `json:"phone"`
-	Password  string `json:"-"`
-	Amount    int    `json:"amount"`
-	Goods     []Good `json:"goods" gorm:"foreignKey:UserID"`
+	FirstName string  `json:"firstName"`
+	LastName  string  `json:"lastName"`
+	Email     string  `json:"email"`
+	Age       string  `json:"age"`
+	Phone     string  `json:"phone"`
+	Password  string  `json:"-"`
+	Amount    float64 `json:"amount"`
+	Goods     []Good  `json:"goods" gorm:"foreignKey:UserID"`
 }
